+++ orcid-core/src/main/java/org/orcid/core/manager/v3/WorksExtendedCacheManager.java
@@ -1,9 +1,13 @@
 package org.orcid.core.manager.v3;
 
+import org.orcid.pojo.WorkSummaryExtended;
 import org.orcid.pojo.WorksExtended;
+
+import java.util.List;
 
 public interface WorksExtendedCacheManager {
 
     WorksExtended getGroupedWorksExtended(String orcid);
 
+    List<WorkSummaryExtended> getFeaturedGroupedWorksExtended(String orcid);
 }