+++ orcid-core/src/main/java/org/orcid/core/manager/v3/read_only/impl/WorkManagerReadOnlyImpl.java
@@ -22,6 +22,7 @@ import org.orcid.core.manager.ClientDetailsEntityCacheManager;
 import org.orcid.core.manager.SourceNameCacheManager;
 import org.orcid.core.manager.WorkEntityCacheManager;
 import org.orcid.core.manager.v3.GroupingSuggestionManager;
+import org.orcid.core.manager.v3.WorksExtendedCacheManager;
 import org.orcid.core.manager.v3.read_only.ClientDetailsManagerReadOnly;
 import org.orcid.core.manager.v3.read_only.WorkManagerReadOnly;
 import org.orcid.core.togglz.Features;
@@ -31,6 +32,7 @@ import org.orcid.core.utils.v3.activities.ActivitiesGroupGenerator;
 import org.orcid.core.utils.v3.activities.WorkComparators;
 import org.orcid.core.utils.v3.activities.WorkGroupAndGroupingSuggestionGenerator;
 import org.orcid.jaxb.model.record.bulk.BulkElement;
+import org.orcid.jaxb.model.v3.release.common.PublicationDate;
 import org.orcid.jaxb.model.v3.release.record.ExternalID;
 import org.orcid.jaxb.model.v3.release.record.ExternalIDs;
 import org.orcid.jaxb.model.v3.release.record.GroupAble;
@@ -50,9 +52,14 @@ import org.orcid.pojo.WorkExtended;
 import org.orcid.pojo.WorkGroupExtended;
 import org.orcid.pojo.WorkSummaryExtended;
 import org.orcid.pojo.WorksExtended;
+import org.orcid.pojo.ajaxForm.Date;
 import org.orcid.pojo.ajaxForm.PojoUtil;
+import org.orcid.pojo.ajaxForm.Text;
+import org.orcid.pojo.ajaxForm.WorkForm;
 import org.orcid.pojo.grouping.WorkGroupingSuggestion;
 import org.springframework.beans.factory.annotation.Value;
+
+import static org.orcid.pojo.ajaxForm.PojoUtil.getWorkForm;
 
 public class WorkManagerReadOnlyImpl extends ManagerReadOnlyBaseImpl implements WorkManagerReadOnly {
     
@@ -69,6 +76,9 @@ public class WorkManagerReadOnlyImpl extends ManagerReadOnlyBaseImpl implements 
     protected WorkEntityCacheManager workEntityCacheManager;
 
     private final Integer maxWorksToRead;
+
+    @Resource
+    private WorksExtendedCacheManager worksExtendedCacheManager;
     
     @Resource
     private GroupingSuggestionManager groupingSuggestionsManager;
@@ -191,18 +201,17 @@ public class WorkManagerReadOnlyImpl extends ManagerReadOnlyBaseImpl implements 
      * Get the list of works that belongs to a user
      *
      * @param orcid
+     * @param featuredOnly
      * @return the list of works that belongs to this user
      */
     @Override
-    public List<WorkSummaryExtended> getWorksSummaryExtendedList(String orcid) {
-        List<WorkSummaryExtended> wseList = retrieveWorkSummaryExtended(orcid);
+    public List<WorkSummaryExtended> getWorksSummaryExtendedList(String orcid, boolean featuredOnly) {
+        List<WorkSummaryExtended> wseList = retrieveWorkSummaryExtended(orcid, featuredOnly);
         // Filter the contributors list
         for (WorkSummaryExtended wse : wseList) {
             if (wse.getContributorsGroupedByOrcid() != null && wse.getContributorsGroupedByOrcid().size() > 0) {
-                contributorUtils.filterContributorsGroupedByOrcidPrivateData(wse.getContributorsGroupedByOrcid(), maxContributorsForUI);
                 wse.setNumberOfContributors(wse.getContributorsGroupedByOrcid().size());
             } else {
-                contributorUtils.filterContributorPrivateData(wse.getContributors().getContributor(), maxContributorsForUI);
                 List<ContributorsRolesAndSequences> contributorsGroupedByOrcid = contributorUtils.getContributorsGroupedByOrcid(wse.getContributors().getContributor(), maxContributorsForUI);
                 wse.setContributorsGroupedByOrcid(contributorsGroupedByOrcid);
                 wse.setNumberOfContributors(contributorsGroupedByOrcid.size());
@@ -211,9 +220,9 @@ public class WorkManagerReadOnlyImpl extends ManagerReadOnlyBaseImpl implements 
         return wseList;
     }
 
-    private List<WorkSummaryExtended> retrieveWorkSummaryExtended(String orcid) {
+    private List<WorkSummaryExtended> retrieveWorkSummaryExtended(String orcid, boolean featuredOnly) {
         List<WorkSummaryExtended> workSummaryExtendedList = new ArrayList<>();
-        List<Object[]> list = workDao.getWorksByOrcid(orcid);
+        List<Object[]> list = workDao.getWorksByOrcid(orcid, featuredOnly);
         for(Object[] q1 : list){
             BigInteger putCode = (BigInteger) q1[0];
             String workType = isEmpty(q1[1]);
@@ -232,6 +241,7 @@ public class WorkManagerReadOnlyImpl extends ManagerReadOnlyBaseImpl implements 
             Timestamp createdDate = (Timestamp) q1[14];
             Timestamp lastModifiedDate = (Timestamp) q1[15];
             String contributors = isEmpty(q1[16]);
+            int featuredDisplayIndex = (int) q1[17];
             ExternalIDs externalIDs = null;
             if (externalIdsJson != null) {
                 externalIDs = jsonWorkExternalIdentifiersConverterV3.convertFrom(externalIdsJson,null);
@@ -272,6 +282,7 @@ public class WorkManagerReadOnlyImpl extends ManagerReadOnlyBaseImpl implements 
                     .sourceName(sourceName)
                     .assertionOriginName(assertionOriginName)
                     .displayIndex(displayIndex)
+                    .featuredDisplayIndex(featuredDisplayIndex)
                     .assertionOriginSourceId(assertionOriginSourceId)
                     .assertionOriginClientSourceId(assertionOriginClientSourceId)
                     .contributors(contributorList)
@@ -385,7 +396,24 @@ public class WorkManagerReadOnlyImpl extends ManagerReadOnlyBaseImpl implements 
 
     @Override
     public WorksExtended getWorksExtendedAsGroups(String orcid) {
-        return groupWorksExtendedAndGenerateGroupingSuggestions(getWorksSummaryExtendedList(orcid), orcid);
+        return groupWorksExtendedAndGenerateGroupingSuggestions(getWorksSummaryExtendedList(orcid, false), orcid);
+    }
+
+    @Override
+    public List<WorkSummaryExtended> getFeaturedWorksSummaryExtended(String orcid) {
+        return getWorksSummaryExtendedList(orcid, true);
+    }
+
+    @Override
+    public List<WorkForm> getFeaturedWorks(String orcid) {
+        List<WorkSummaryExtended> works = worksExtendedCacheManager.getFeaturedGroupedWorksExtended(orcid);
+        List<WorkForm> workForms = new ArrayList<>();
+        if (!works.isEmpty()) {
+            for (WorkSummaryExtended workSummary : works) {
+                workForms.add(getWorkForm(workSummary));
+            }
+        }
+        return workForms;
     }
 
     private String[] getPutCodeArray(String putCodesAsString) {