+++ orcid-core/src/main/java/org/orcid/core/manager/v3/validator/ActivityValidator.java
@@ -21,6 +21,8 @@ import org.orcid.core.exception.ActivityTypeValidationException;
 import org.orcid.core.exception.InvalidAmountException;
 import org.orcid.core.exception.InvalidDisambiguatedOrgException;
 import org.orcid.core.exception.InvalidFuzzyDateException;
+import org.orcid.core.exception.InvalidOrgAddressCityNoCountryException;
+import org.orcid.core.exception.InvalidOrgAddressException;
 import org.orcid.core.exception.InvalidOrgException;
 import org.orcid.core.exception.InvalidPutCodeException;
 import org.orcid.core.exception.MissingStartDateException;
@@ -55,6 +57,9 @@ import org.orcid.jaxb.model.v3.release.common.Visibility;
 import org.orcid.jaxb.model.v3.release.common.Year;
 import org.orcid.jaxb.model.v3.release.groupid.GroupIdRecord;
 import org.orcid.jaxb.model.v3.release.record.Affiliation;
+import org.orcid.jaxb.model.v3.release.record.AffiliationType;
+import org.orcid.jaxb.model.v3.release.record.Education;
+import org.orcid.jaxb.model.v3.release.record.Employment;
 import org.orcid.jaxb.model.v3.release.record.ExternalID;
 import org.orcid.jaxb.model.v3.release.record.ExternalIDs;
 import org.orcid.jaxb.model.v3.release.record.ExternalIdentifiersAwareActivity;
@@ -374,6 +379,31 @@ public class ActivityValidator {
             throw new InvalidDisambiguatedOrgException();
         }
     }
+    
+    
+    private void validateOrgAddress(OrganizationHolder organizationHolder) {
+        if (organizationHolder.getOrganization() == null) {
+            throw new InvalidOrgException();
+        }
+
+        Organization org = organizationHolder.getOrganization();
+        if (org.getAddress() == null || PojoUtil.isEmpty(org.getAddress().getCity()) || org.getAddress().getCountry() == null
+                || PojoUtil.isEmpty(org.getAddress().getCountry().name())) {
+            throw new InvalidOrgAddressException();
+        }
+    }
+    
+    private void validateOrgAddressForProfessionalActivities(OrganizationHolder organizationHolder) {
+        if (organizationHolder.getOrganization() == null) {
+            throw new InvalidOrgException();
+        }
+
+        Organization org = organizationHolder.getOrganization();
+        if (org.getAddress() != null && !PojoUtil.isEmpty(org.getAddress().getCity()) && ( org.getAddress().getCountry() == null
+                || PojoUtil.isEmpty(org.getAddress().getCountry().name()))) {
+            throw new InvalidOrgAddressCityNoCountryException();
+        }
+    }
 
     private void validateDisambiguatedOrg(MultipleOrganizationHolder organizationHolder) {
         if (organizationHolder.getOrganization() == null) {
@@ -402,9 +432,18 @@ public class ActivityValidator {
             Visibility updatedVisibility = affiliation.getVisibility();
             validateVisibilityDoesntChange(updatedVisibility, originalVisibility);
         }
+        
 
         if (isApiRequest) {
             validateDisambiguatedOrg(affiliation);
+            //validate city/country requirement for education and employment 
+            if (affiliation instanceof Education || affiliation instanceof Employment) {
+                validateOrgAddress(affiliation);
+            }
+            else {
+                validateOrgAddressForProfessionalActivities(affiliation);
+            }
+
             if (affiliation.getEndDate() != null) {
                 validateFuzzyDate(affiliation.getEndDate());
             }