+++ orcid-core/src/main/java/org/orcid/core/togglz/Features.java
@@ -71,7 +71,13 @@ public enum Features implements Feature {
     ENABLE_PAPI_RATE_LIMITING,
 
     @Label("Use the authorization server to sign in")
-    OAUTH_SIGNIN;
+    OAUTH_SIGNIN,
+
+    @Label("Use the authorization server to authorize OAUTH requests")
+    OAUTH_AUTHORIZATION,
+
+    @Label("Enable featured works logic in the UI")
+    FEATURED_WORKS_UI;
     
     public boolean isActive() {
         return FeatureContext.getFeatureManager().isActive(this);