+++ orcid-core/src/test/java/org/orcid/core/manager/v3/WorkManagerTest.java
@@ -92,8 +92,7 @@ import org.orcid.persistence.jpa.entities.ClientDetailsEntity;
 import org.orcid.persistence.jpa.entities.MinimizedWorkEntity;
 import org.orcid.persistence.jpa.entities.PublicationDateEntity;
 import org.orcid.persistence.jpa.entities.WorkEntity;
-import org.orcid.pojo.ContributorsRolesAndSequences;
-import org.orcid.pojo.WorkExtended;
+import org.orcid.pojo.*;
 import org.orcid.pojo.ajaxForm.Text;
 import org.orcid.pojo.ajaxForm.WorkForm;
 import org.springframework.beans.factory.annotation.Value;
@@ -1971,6 +1970,19 @@ public class WorkManagerTest extends BaseTest {
         Mockito.verify(mockNotificationManager, Mockito.times(2)).sendAmendEmail(any(), any(), any());
 
         workManager.removeWorks(claimedOrcid, Arrays.asList(work.getPutCode()));
+    }
+
+    @Test
+    public void a_testGetWorksExtendedAsGroups() {
+        WorksExtended works = workManager.getWorksExtendedAsGroups(claimedOrcid);
+        List<WorkGroupExtended> workGroup = works.getWorkGroup();
+        assertEquals(3, workGroup.size());
+    }
+
+    @Test
+    public void a_testGetFeaturedWorksSummaryExtended() {
+        List<WorkSummaryExtended> works = workManager.getFeaturedWorksSummaryExtended(claimedOrcid);
+        assertEquals(2, works.size());
     }
 
     private WorkEntity getUserPreferredWork() {