+++ orcid-scheduler-web/src/main/java/org/orcid/scheduler/mvp/OrcidIntegrationMVPNotifications.java
@@ -8,13 +8,17 @@ import javax.annotation.Resource;
 
 import org.apache.commons.lang3.StringUtils;
 import org.json.JSONArray;
+import org.orcid.core.manager.ProfileEntityCacheManager;
 import org.orcid.core.manager.v3.NotificationManager;
+import org.orcid.core.manager.v3.RecordNameManager;
+import org.orcid.jaxb.model.v3.release.record.Name;
 import org.orcid.persistence.dao.ClientDetailsDao;
 import org.orcid.persistence.dao.OrcidOauth2TokenDetailDao;
 import org.orcid.persistence.dao.ProfileEmailDomainDao;
 import org.orcid.persistence.jpa.entities.ClientDetailsEntity;
 import org.orcid.persistence.jpa.entities.NotificationEntity;
 import org.orcid.persistence.jpa.entities.ProfileEmailDomainEntity;
+import org.orcid.persistence.jpa.entities.ProfileEntity;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -41,6 +45,19 @@ public class OrcidIntegrationMVPNotifications {
     @Resource(name = "orcidOauth2TokenDetailDaoReadOnly")
     private OrcidOauth2TokenDetailDao orcidOauth2TokenDetailDaoReadOnly;
 
+    @Resource(name = "recordNameManagerV3")
+    private RecordNameManager recordNameManager;
+
+    /**
+     * Creates notifications for clients that have enabled the MVP feature
+     * 
+     * This method checks for clients that have the MVP feature enabled and
+     * sends notifications to users whose email domains match the configured
+     * domains for each client. Notifications are sent only if no previous
+     * notifications have been sent or no notification sent within a specified
+     * number of days by default 10.
+     */
+
     public void createOrcidIntegrationNotifications() {
         // Get clients eligible for mvp
         long startTime = System.currentTimeMillis();
@@ -50,6 +67,16 @@ public class OrcidIntegrationMVPNotifications {
             for (ClientDetailsEntity clientDetails : clientsWithMVP) {
                 if (StringUtils.isNotBlank(clientDetails.getNotificationWebpageUrl()) && StringUtils.isNotBlank(clientDetails.getNotificationDomains())) {
                     long startTimeClient = System.currentTimeMillis();
+                    Name memberNameObj = recordNameManager.getRecordName(clientDetails.getGroupProfileId());
+                    String memberName = null;
+                    if (memberNameObj != null) {
+                        memberName = memberNameObj.getCreditName() != null ? memberNameObj.getCreditName().getContent() : null;
+                    }
+
+                    if (memberName == null) {
+                        memberName = clientDetails.getClientName();
+                    }
+
                     LOG.info("Start process client {}. Notification send for the second time if no notifications sent since {} ago", clientDetails.getClientId(),
                             daysAgo);
                     try {
@@ -63,7 +90,7 @@ public class OrcidIntegrationMVPNotifications {
                             }
                         }
                         if (profileDomainSet.size() > 0) {
-                            
+
                             for (ProfileEmailDomainEntity pe : profileDomainSet) {
                                 if (!orcidOauth2TokenDetailDaoReadOnly.hasTokenForClient(pe.getOrcid(), clientDetails.getClientId())) {
                                     List<NotificationEntity> orcidIntegrationNotifications = notificationManager.findByOrcidAndClientAndNotificationFamilyNoClientToken(
@@ -71,7 +98,7 @@ public class OrcidIntegrationMVPNotifications {
                                     boolean shouldSendNotification = orcidIntegrationNotifications == null || orcidIntegrationNotifications.isEmpty()
                                             || shouldSendNotification(orcidIntegrationNotifications);
                                     if (shouldSendNotification) {
-                                        notificationManager.sendOrcidIntegrationNotificationToUser(pe.getOrcid(), clientDetails);
+                                        notificationManager.sendOrcidIntegrationNotificationToUser(pe.getOrcid(), clientDetails, memberName);
                                         LOG.warn("Create notification for client {} and orcid {}", clientDetails.getClientId(), pe.getOrcid());
                                     } else {
                                         LOG.info("Notification already sent for client {} and orcid {}", clientDetails.getClientId(), pe.getOrcid());
@@ -107,19 +134,17 @@ public class OrcidIntegrationMVPNotifications {
         long seconds = (durationMillis / 1000) % 60;
         return String.format("%d hours, %d minutes, %d seconds", hours, minutes, seconds);
     }
-    
-    
+
     private boolean shouldSendNotification(List<NotificationEntity> orcidIntegrationNotifications) {
         if (orcidIntegrationNotifications == null || orcidIntegrationNotifications.isEmpty()) {
             return true;
         }
-        
-        if (orcidIntegrationNotifications.size() > 1)   {
-            return false ;
-        }
-        else {
-        java.util.Date daysAgoDate = new java.util.Date(System.currentTimeMillis() - daysAgo * 24 * 60 * 60 * 1000);
-        return orcidIntegrationNotifications.get(0).getDateCreated() != null && orcidIntegrationNotifications.get(0).getDateCreated().before(daysAgoDate);
+
+        if (orcidIntegrationNotifications.size() > 1) {
+            return false;
+        } else {
+            java.util.Date daysAgoDate = new java.util.Date(System.currentTimeMillis() - daysAgo * 24 * 60 * 60 * 1000);
+            return orcidIntegrationNotifications.get(0).getDateCreated() != null && orcidIntegrationNotifications.get(0).getDateCreated().before(daysAgoDate);
         }
     }
 }